requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
requests
faiss-cpu
numpy
twilio
mistralai
python-multipart